fix zsh-autoquoter when loaded before zsh-syntax-highlighter
If you tried to load zsh-autoquoter before zsh-syntax-highlighter, the
assignment to ZSH_HIGHLIGHT_STYLES[zaq:string] would fail because
ZSH_HIGHLIGHT_STYLES was not known to be an associative array. It would
manifest as this error:

    zsh-autoquoter.zsh:85: bad math expression: ':' without '?'

Now zsh-autoquoter will forward-declare it if it isn't set yet, so that the two
plugins can be loaded in any order, and so that you can use zsh-autoquoter
without using zsh-syntax-highlighter.